requirements.txt
flask==3.1.1
numpy==2.3.0
gunicorn==23.0.0
python-dotenv
Flask-Cors==6.0.1
pandas==2.3.0
google-generativeai==0.8.5
python-dotenv==1.1.0
google==3.0.0
google-generativeai==0.8.5